memoryuser: don't cleanup dead crostini allocator
multivm.GameMode.crostini failed because we couldn't clean up the
allocator process, because it was already dead. So only clean it up if
it's not dead.

BUG=b:246896386
TEST=tast multivm.GameMode.crostini

+++ src/chromiumos/tast/local/memory/memoryuser/crostini_lifecycle_task.go
@@ -82,6 +82,10 @@ func (t *CrostiniLifecycleUnit) Close(ctx context.Context) error {
 	if t.cmd == nil {
 		return nil
 	}
+	if t.cmd.ProcessState == nil {
+		return nil
+	}
+	// Only kill the allocator if it is running.
 	if err := t.cmd.Kill(); err != nil {
 		return errors.Wrap(err, "failed to kill Crostini lifecycle unit")
 	}
